requirements.txt
streamlit
pillow
transformers